requirements.txt
pyaml
python-pptx
openai
nltk
numpy
django
django-cors-headers
certifi
replicate
langchain
tiktoken
torch
safe_import
sympy